api/tacticalrmm/requirements-dev.txt
black
Werkzeug
django-extensions
isort
types-pytz
django-silk
mypy==0.982
django-stubs
pandas-stubs
djangorestframework-stubs
django-types
djangorestframework-types
celery-types
msgpack-types
types-Markdown
types-requests
types-PyYAML
types-urllib3
